backend/pyproject.toml
[project]
name = "bulk-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "fastapi>=0.124.4",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "uvicorn>=0.38.0",
]
